requirements.txt
ollama
chromadb
sentence-transformers
gradio
fpdf2
langchain-community
langchain
openai-whisper
python-dotenv
